// src/types/enums/enums.ts
export enum SortOptions {
    Stars = 'stars',
    Forks = 'forks',
    Updated = 'updated',
    HelpWantedIssues = 'help-wanted-issues',
  }

export enum ColumnHeaders{
    Repository = 'Repository',
    Description = 'Description',
    Stars = 'Stars',
    Forks = 'Forks',
    Language = 'Languages',
    Updated = 'Updated At',
    Topics = 'Topics',
}